Fixes #4753: Use the vanilla helper for SUCCESS or CONSUME (#5072)
This makes now use of the vanilla helper to return `SUCCESS` or
`CONSUME` depending on  the side.

Also some general cleanup like unused imports, missing file headers, etc

+++ src/main/java/appeng/parts/automation/FormationPlanePart.java
@@ -438,7 +438,7 @@ public class FormationPlanePart extends AbstractFormationPlanePart<IAEItemStack>
 
         @Override
         public float getPlacementYaw() {
-            return (float) (this.lookDirection.getHorizontalIndex() * 90);
+            return this.lookDirection.getHorizontalIndex() * 90;
         }
     }
 }
